requirements.txt
Django==1.9.1
djangorestframework==3.3.2
gunicorn==19.4.5
mysqlclient==1.3.7
Pillow==3.1.0
wheel==0.26.0
